🏗️ Refactor(PriceList.jsx): 폴더 구조 변경
Payment -> Price로 이동
