+++ react-app/src/components/Price/Price.jsx
