+++ react-app/src/pages/PriceList.jsx
@@ -1,4 +1,4 @@
-import Price from '../components/Payment/Price';
+import Price from '../components/Price/Price';
 import Button from '../components/Button/Button';
 import { useNavigate } from 'react-router-dom';
